Set overlayfile as kwarg & vectorise masking procedure
Overlayfiles used to be an argument that had to be given, but it's reasonable to want to plot a surface without any overlay. You can now not pass any overlayfiles.

Also added `display` to the widget; this means that the widget is always shown and you don't need to call widget.fig in the notebook